SoftValue is a lot softer now; new strategy for pre-request Stratum activity; better performance in ActionMatcher.

+++ EcosystemSimulation/src/shmp/utils/MovingAverage.kt
@@ -3,13 +3,13 @@ package shmp.utils
 
 data class MovingAverage(var value: SoftValue, val coefficient: Double) : Comparable<MovingAverage> {
     constructor(double: Double, coefficient: Double): this(SoftValue(double), coefficient)
-    constructor(int: Int, coefficient: Double): this(SoftValue(int.toDouble()), coefficient)
+    constructor(int: Int, coefficient: Double): this(SoftValue(int), coefficient)
 
     fun change(new: SoftValue) {
         value = coefficient * value + (1 - coefficient) * new
     }
     fun change(double: Double) = change(SoftValue(double))
-    fun change(int: Int) = change(SoftValue(int.toDouble()))
+    fun change(int: Int) = change(SoftValue(int))
 
     override fun toString() = value.toString()
